restore the "print ASAP" feature
Since the DB was introduced, computation of hashes is no more lazy,
and the program was then computing all the hashes before starting
to print any result.
Note that with the option to sort on sizes, everything must still be
computed before printing starts (unlike before the DB). Changing that
requires too much complexity in the code for seemingly too little gain.