use RawFilePath (phase 1)
The main motivation is that we use file paths as "data", in particular
we need to hash them, irregardless of any encoding.
This BTW uncovered a bug in `combine`, were the conversion to ByteString
was using ByteString.Char8.pack, which can discard data.
Some significant speed-up (e.g. ~ 2x) can be observed in some cases.

This is just a first pass, and leaves the program in a more incorrect state
than before w.r.t. encoding.